// packages/shared-components/Select/types.ts
export type OptionShape = {
    /**
     * Текстовое представление пункта
     */
    key: string;

    /**
     * Контент, который будет отрисован в выпадающем списке и в поле при выборе
     */
    content?: ReactNode;

    /**
     * Блокирует данный пункт для выбора
     */
    disabled?: boolean;

    /**
     * Разрешает показ компонента Checkmark, иногда нужно его убирать для показа контента ошибки или пустого состояния
     */
    showCheckMark?: boolean;

    /**
     * Дополнительные данные
     */
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    value?: any;

    // Опция-прелоадер
    isPreloader?: boolean;
};
// ...
export type OptionProps = {
    /**
     * Дополнительный класс
     */
    className?: string;

    /**
     * Тема
     */
    isMobile?: boolean;

    /**
     * Контент пункта меню
     */
    children?: ReactNode;

    /**
     * Данные пункта меню
     */
    option: OptionShape;

    /**
     * Индекс пункта
     */
    index: number;

    /**
     * Флаг, выбран ли данный пункт
     */
    selected?: boolean;

    /**
     * Флаг, подсвечен ли данный пункт
     */
    highlighted?: boolean;

    /**
     * Флаг, заблокирован ли данный пункт
     */
    disabled?: boolean;

    /**
     * Флаг множественного выбора
     */
    multiple?: boolean;

    /**
     * Компонент пункта меню
     */
    Checkmark?: FC<CheckmarkProps> | null;

    /**
     * Внутренние свойства, которые должны быть установлены компоненту.
     */
    innerProps: {
        id: string;
        onClick: (event: MouseEvent<HTMLLIElement>) => void;
        onMouseDown: (event: MouseEvent<HTMLLIElement>) => void;
        onMouseMove: (event: MouseEvent<HTMLLIElement>) => void;
        role: string;
    } & RefAttributes<HTMLLIElement> &
        AriaAttributes;
};

export type CheckmarkProps = {
    /**
     * Флаг, данный пункт выбран
     */
    selected?: boolean;

    /**
     * Флаг, данный пункт задизейблен
     */
    disabled?: boolean;

    /**
     * Дополнительный класс
     */
    className?: string;

    /**
     * Флаг множественного выбора
     */
    multiple?: boolean;

    /**
     * Расположение отметки
     */
    position?: 'before' | 'after';

    /**
     * Иконка выбранного пункта
     */
    icon?: FC<SVGProps<SVGSVGElement>>;

    checkboxCSS?: CSSObject;
};